create/requirements.txt
Click==7.0
flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
google-cloud-datastore==2.13.2